+++ Link/extensions.cpp
@@ -1,0 +1,8 @@
+/** @file
+ * Contains CPU and multithreaded specific high-performance backend facilities,
+ * which can be considered as extensions to the QuEST API.
+ *
+ * @author Tyson Jones
+ */
+ 
+# include "QuEST.h"
